first commit-- categorized visualization script and crawling script
Twitter is a micro-blogging website and it's growing exponentially .
Performed sentimental analysis of tweet and analysed the hot-topic
"Demonetisation in India", whether people will support or boycott it, by
collecting a large set of tweet. Visualization is worth more than words,
so visualized the said analysis in a neat-nice way to make people
understand more easily.